pyproject.toml
[tool.poetry]
name = "trueformtranslator"
version = "0.1.0"
description = ""
authors = ["sertemo <tejedor.moreno@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.29.0"
langchain = "^0.1.0"
langdetect = "^1.0.9"
pycountry = "^23.12.11"
textblob = "^0.17.1"
python-docx = "^1.1.0"
scikit-learn = "^1.3.2"
tiktoken = "^0.5.2"
openai = "^1.7.0"
cohere = "^4.41"
langchain-openai = "^0.0.2"
langchain-community = "^0.0.11"
langchain-core = "^0.1.9"
python-dotenv = "^1.0.0"
pandas = "^2.1.4"
pydantic = "^2.5.3"
pymongo = "^4.6.1"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
